@@ -1,21 +1,21 @@
 module github.com/gostaticanalysis/knife
 
-go 1.22.2
+go 1.23.4
 
 require (
 	github.com/gostaticanalysis/analysisutil v0.7.1
 	github.com/gostaticanalysis/astquery v0.0.0-20200823120951-321f091076cd
-	github.com/gostaticanalysis/comment v1.4.2
-	golang.org/x/tools v0.20.0
+	github.com/gostaticanalysis/comment v1.5.0
+	golang.org/x/tools v0.28.0
 )
 
 require (
-	github.com/antchfx/xpath v1.3.0 // indirect
+	github.com/antchfx/xpath v1.3.2 // indirect
 	github.com/google/go-cmp v0.6.0 // indirect
 	github.com/gostaticanalysis/testutil v0.4.0 // indirect
 	github.com/hashicorp/go-version v1.6.0 // indirect
 	github.com/otiai10/copy v1.14.0 // indirect
-	golang.org/x/mod v0.17.0 // indirect
-	golang.org/x/sync v0.7.0 // indirect
+	golang.org/x/mod v0.22.0 // indirect
+	golang.org/x/sync v0.10.0 // indirect
 	golang.org/x/text v0.14.0 // indirect
 )
